Fixed the file paths error
