--- BufferData\CurrentUser\username.txt
@@ -1,1 +1,0 @@
-hello
