--- bus-booking-system/BufferData/CurrentUser/password.txt
@@ -1,1 +1,0 @@
-1234
