--- bus-booking-system/Modules/DeleteAccount/DeleteTransition.txt
@@ -1,1 +1,0 @@
-True
